+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -54,6 +54,7 @@ Release History
 * `deployment user set`: logged more informative error messages.
 * functionapp: add support for creating Linux function apps
 * appservice: fix list-locations
+* webapp: add --number-of-lines to 'az webapp log tail' currently only supported for linux
 
 0.1.19
 ++++++